requirements.txt
gitdb==4.0.12
GitPython==3.1.44
python-dateutil==2.9.0.post0
six==1.17.0
smmap==5.0.2
